requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cloudinary==1.39.0
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.7.1
dj-database-url==2.1.0
Django==5.0.3
django-allauth==0.61.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-jazzmin==2.6.0
djangorestframework==3.15.1
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
oauthlib==3.2.2
packaging==24.0
pillow==10.2.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
pypng==0.20220715.0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.1.0
requests==2.31.0
requests-oauthlib==1.4.0
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
